Reorganized the tiles in the XPM files.  The 15x20 tiles are now in small.xpm, the thumbs and "diplomatic treaty" tiles are in treaty.xpm.
The server will figure out the size of the tiles from the size of the XPM.

Fixed some graphic "bugs" in the tiles.  Changed "none" to "None" so GIMP
and XV would understand the XPM files.

Added two new tiles, a small + to indicate stacked units, and a cross-hair
type cursor for indicating goto destinations.

Added code so the + is put on unit stacks.  This is how real Civ works, you
get to know the difference between 1 unit and 2+ units on a tile.