dwp@mso.anu.edu.au: Improved "Build City" menu item and related issues.
Patches submitted by Jeff Mallatt <jjm@codewell.com>:

When a Settler/Engineer is sitting on a City, pressing 'b' will
order it to add itself to that City (if the City is small enough).
However, the Orders menu Build City item remains disabled.
This patch enables the menu item in this case, and also renames
it to "Add to City" when appropriate.  The menu item is disabled
if the city is already too big.  This patch also changes
handle_city_info to call update_menus after the City
size has increased due to a Settler being added.

Additional changes by me:

New function can_unit_add_to_city(), which also takes into account
aqueduct_size and sewer_size (for rulesets generality).
In request_unit_build_city(), let the server decide the message if
the build order in inappropriate (to avoid unnecessary code duplication).
In handle_unit_build_city(), did some re-organisation, and adjusted
some of the messages for when the build/add-to order is inapproriate.