We can see intelligence dialog info from someone we're in contact with but not yet embassy with, but we can't see their techs, including what they are currently researching, nor their tax settings.
This means that when you open up the diplomacy dialog to such a player,
you can give your own techs to the other side, and he can put his techs
on the table, but you can't suggest _his_ techs into the proposal,
because you don't know and can't see what techs he has.

This closes #4200 and #4102.