find_the_shortest_path in server/gotohand.c had a static variable local_vector, a 2d array with dimensions MAP_MAX_WIDHT * MAP_MAX_HEIGHT.
This is inefficient not only because of the wasted memory, but because
the inferior cache usage slows things down.

The correct thing to do - and what is done everywhere else - is to make
this value a 1d array indexed by index positions.

Also the array elements are changed from char to bit-vector.
The corresponding BV_ are now used to access the bits.

Noticed and changed by Jason, PR#4636.