Rename the field flags "pre" to "pre-send" and "post" to "post-recv". Add the flag "post-send". Remove the packet type arguments from the callback signature.
Add a new function conn_clear_packet_cache which resets the delta
state of this connection.

Add post-send and post-recv callback for the GAME_STATE packet. This
is necessary since the client removes information cached in map.tiles
when such a game is recevied. However previously this information
(about the removal of the tile data at the client) was not know to the
delta layer. This resulted in not sending the required packets in case
the client connection switches between players.

Reported by Per I. Mathisen in PR#7216. Patch from me.