This patch fixes 3 bugs in settler activities.
1. The biggest bug is that activity times are multiplied by 10. But
settlers.c doesn't know about this. Thus the calculated times to do any
activity are 10x too high (on top of the move time, which is correct).
No doubt this has something to do with the reported problems with
autosettlers.

2. A second bug is in the considering of making a road plus a rail in
settlers.c. There is a can_rr variable that is used to see if the
settler can make a railroad. But it's not used to see if the settler
can make a road plus a railroad. The fix is easy.

3. The third bug is that in unittools.c when adding on to the
activity_count there is a special case for when the unit has nove move
points left. This is a bug because the amount added on has nothing to do
with the MP. So an engineer with 1/3 MP left gets the same amount of
work done as if it had 2 MP, but with 0 MP left it only gets half this
amount.

Added a new function get_turns_for_activity (equivalent to the function
needed by gui-ftwl). Moved get_settler_speed into unit.c and changed it
to do all the work of calculating the speed (instead of just half of the
work).

Patch by Jason in PR#9561.