Do not attempt to become a unit bodyguard when assigned to guard duty for a city.
Do not yoyo between cities attempting to defend them all.
look_for_charge() has the problem that it calculates current defense
in each city in the world, but forgets to calculate in that once we
leave the city we are currently in, that city might become the most
needy! So instead we try harder to stay where we are.

When military code can't find us work to do, actually do go home.
Now using ai_military_defend().