Add AI reactions to diplomat incidents, declarations of war and nuclear explosions, even if not done against that player. This affects 'AI love'.
Also changes Senate so that it will only prevent declarations of war, not every treaty
reduction.

For declarations of war, we will remember the highest level of treaties between players
ever, and punish AI attitudes towards war between former allies and friends, which is an
indication of backstabbing and betrayal. Patch in PR#13361.