(PR#39383) BUG: terrain and resource name translation
Tests were wildly inconsistent, sometimes testing against ruleset name string, 
others against translated name.

Renamed name_orig to name_rule (the origin), and name to name_translated.  
Translated names are now only accessed via a central function.  (The function 
existed, but not everybody used it.)

Moved the translation into the accessor functions, as this seemed to be better 
than spread around in 3 places, especially as the other places in the code 
forgot to test for NULL and '\0' (empty string).