(PR#39424) BUG: improvement name translation
Fixed several tests against ruleset name strings.

Renamed name_orig to name_rule (the origin), and name to name_translated.  
Translated names are now only accessed via a central function.  (The function 
existed, but not everybody used it.)

Moved the translation into the accessor functions, as this seemed to be better 
than spread around in 3 places, especially as the other places in the code 
forgot to test for NULL and '\0' (empty string).

Plus the usual cleanup of translated names for untranslated log messages.