(PR#39451) Add signal emit to scripting api
Make sure the scripting api action give_technology sends a 
script signal message for every researched technology.
give_technology now takes a new argument reason that is sent
with the signal, and it should almost always be "hut" for
scripts.

It was decided that script signal emission should not be 
exposed to the scripting api.