API: Define convenience functionality for Types
Define the == equality operator for all API datatypes, so you can
compare objects directly (instead of comparing the id field, which is
of course what the equality operation does).

Define a string representation, so that when printed, userdata types
output something like the following (instead of userdata: 0x0x1080ac48):

  <Government #2 Monarchy>
  <Unit #114>
  <Player #0 Agamemnon>
  <City #137 Agadir>

A single quote (') is removed in a comment, it disturbs syntax
highlighting if you want to view api.pkg in Lua syntax colors.

See gna patch #1562