Increase the height of the info box in the happines tab in the city dialogue
The happiness tab of the city dialogue is taller than the available area
on my my netbook's screen, which forces the last few lines of the info
box to be hidden behind the happiness box.

This is not critical as all information is available in the overview
tab, but as the happiness box does not need the entire width of the
window, I rearranged the widgets and is attaching a patch here.

patch by Jon Severinsson (jon_severinsson)

see gna patch #1711