add check to tile_transform
This check is needed as tile could be a virtual tile used by the client
(gtk) to check which unit tasks should be shown. If a city is on the
tile and it could be transformed to a terrain with the flag
TER_NO_CITIES this would result in an assert().

see gna bug #18549