Introduce meta knowledge and use it when evaluating action enablers in the GTK3 client.
The server knows everything. The player has limited knowledge. It is
possible to create rules that depend on knowledge the player don't have.
This leads to situations where the player don't know if a rule applies or
not. Meta knowledge helps identify those situations.

This patch introduces meta knowledge about requirements in requirement
vectors. It is not complete. What is not explicitly said to be certain is
assumed to be uncertain. It lives in common so a hypothetical future non
cheating server side AI can use it.

The meta knowledge introduced is used to evaluate action enablers in the
GTK3 client. This evaluation determine how it is showed to the user. When it
is uncertain if an action is enabled or not it is displayed but colored red.
Nothing changes when the client is certain.

Meta knowledge about the requirements of the action enablers used in the
default rule sets are added.

See bug #21105