Make trying to do an illegal action cost a bit movement.
When an action enabler depends on game state the player don't know an action
being refused may make it possible to infer the unknown game state. In the
case of actions that are impossible to cancel once requested this
information isn't free since the action will be done if it is legal. Make
trying an illegal action cost a bit movement to make sure the information
always has a cost.

Thanks to Marko Lindqvist <cazfi> and Jacob Nevins <jtn> for feed back on
this issue.

See patch #4199