Don't battle your own on the behalf of the ally you spy on.
If a unit that has the Diplomat flag or the SuperSpy flag is on the same
tile as the target of certain diplomat actions it will try to stop them via
diplomatic combat. A unit should not try to stop an action done by it self
or another unit controlled by the same player.

See bug #21596