Use tile ID when asking user to choose diplomatic action
Change the target information from victim ID to tile ID in the client's
diplomat_arrival_queue and in the server's request for a clarification about
what the player intended by moving an unit with the Diplomat flag. Since the
client didn't send the server supplied potential victim ID to
popup_diplomat_dialog() no server control is lost.

When moving an unit that has the Diplomat flag it can be unclear to the
server if a move or a diplomatic action was intended. The server must
therefore request this information from the client. When the server request
this information, or the player request a diplomatic action, the client
don't know what action the player asked for. The client must therefore ask
the player. Since the player only can answer one question at a time the
questions are queued. Game state related to a queued question may change
while the player replies to a previous question.

The server used to ask by sending the client the ID of the Diplomat and the
ID of a potential victim on the tile. The client would then put this
information in the queue of questions. The potential victim may move before
the diplomat action dialog is shown to the player. If the player intended
to move to the tile it self or target another victim on the tile this
behavior gave an unexpected result.

See patch #4476