AI: consider how a building affects a city as a potential action target
Take any significant change in what actions a city can be the target of into
account when deciding how much the city should value a building. To be seen as
significant the change must be between never possible and sometimes possible.

All actions that currently are controlled by action enablers are negative for
the target city. The value of being a target of an action is therefore given as
negative utility. ("How bad is it?")

The negative utility of being a potential victim of Incite City is taken from
the No_Incite effect. The negative utility of being a potential victim of the
other actions are set to be low (max 10) but with enough granularity to make the
AI prioritize what actions to block.

See patch #4622