Infer if a unit type is an actor from the action enablers, not Diplomat
Up until now a unit have been seen as a potential actor if it has the
Diplomat unit type flag. It is possible to have a unit with the Diplomat
unit type that can't perform any actions. In the future it should be
possible to have actors without the Diplomat unit type flag.

Stop relying on the Diplomat unit flag to determine if the unit type can
perform generalized actions. Check if at least one action controlled by
generalized (ruleset) action enablers is possible in stead.

See patch #5102