Use the knowledge of the player to see if a move triggers an action.
The server side function unit_move_handling() figures out if a move triggers
a spy action. If it triggers a spy action the client is notified. This was
done using full server side knowledge. A spy action may be enabled by a rule
that depends on knowledge the player don't have. If all actions are illegal
based on knowledge the player don't have the client should still be noticed.
Use the knowledge of the player to see if an action was triggered in stead
so this happens.

See bug #22561