Add a lua signal for action started for each actor/target kind pair in use.
The signals are action_started_unit_city (for a unit acting on a city) and
action_started_unit_unit (for a unit acting on another unit). Their only
difference is the kind of action that trigger them and the kind of target in
their target parameter. They are emitted right before an action is run, but
after it is known to be legal. Should false flag support later be added
their information is still intended to be the truth.

The signals have three parameters. The first parameter is the action that is
about to be started. The second is the actor. In both the actor is a unit.
The third parameter is the target. In action_started_unit_city it is a city.
In action_started_unit_unit it is a unit.

See patch #5240