Don't transfer dying units to other players.
Transferring a dying GameLoss unit will kill the player he is transferred to
rather than the player who owned it when it was killed.

I think this happened to the reporter of bug #23676. The gameloss_style was
Loot. A Leader (a GameLoss unit) was inside a city. The Leader was killed.
The city where the Leader was located was transferred to his killer as part
of the Loot.

If the dying Leader is transferred with the city he is inside the killer
will loose the game while the player the Leader used to belong to lives on
without the vulnerability of having a Leader.

Reported anonymously

See bug #23676