Save game unit order upgrade skeleton
Some unit order types can be expressed as a perform action order. To remove
duplication a order like that should be replaced with the corresponding
perform action order. This will require that the save game loading updates
the orders in a savegame from a previous Freeciv release.

Create a skeleton for unit order save game compatibility. No orders are
updated yet. The update can't be a simple text change because what action an
order maps to depends on data like if the target tile has a city.

See patch #6186