Handle ACTION_ANY in move frags may act code
Handle the special value ACTION_ANY, meaning any action, in
utype_may_act_move_frags(). This is the code responsible for figuring out if
a unit will be able to perform an action if it has a certain amount of move
fragments left. Stop handling it in its users.

See patch #6253