Add a common check if a city is occupied
Where to look when checking if a city is occupied depends on who is doing
it. The server is supposed to check units at the tile. The client has the
occupied city property.

This can cause copy past errors.

Add the new common function city_is_occupied(). It will check its context
before looking in the correct place.

Use it in the popup info text generation so Freeciv-web won't start claiming
that occupied cities are unoccupied if it is copied again.

Reported by Nook in his Let's Play Freeciv Youtube video at
https://youtu.be/NzJhIDeVEnE?t=29m21s

See bug #23813