Close wonder specific "Help Wonder" loop hole
The ruleset can now forbid doing an action to a city based on what it
builds. This is less useful for the "Help Wonder" action than expected.

A rule that forbids using a caravan to help a specific wonder can be avoided
in many rulesets. The rule is avoided by switching to a wonder that legally
can received help from a caravan, get the shields and then switch back.
Rulesets where no wonder that can receive caravan shield can be built at the
same time as a wonder that can't aren't affected.

Close the loop hole by only allowing the caravan shields to count if the
new production can be helped by the "Help Wonder" action.

A related loop hole remains for complex rule sets: If "Wonder A" only can
receive help from "Unit 1" while "Wonder B" can receive help from "Unit 2" a
player could switch from "Wonder A" to "Wonder B", use a "Unit 2" to help
build the wonder and then switch back. Solving that isn't in the scope of
this patch.

Only caravan shields are touched. Switching between the hard coded
categories still halves the amount of shields.

See patch #6357