Stricter move as tile action heuristics
A move can be interpreted as an attempt at doing an action. The player is
then showed the action selection dialog where he can choose an action,
cancel or keep moving.

Be more conservative about when a potentially legal action against a tile
will pop up the action selection dialog. It used to pop up if another
target kind existed at the tile. Add the restriction that the actor unit
must be unable to move to the target tile.

See patch #6392