Define the constant DIR8_ORIGIN
Use it, rather than the magic value -1, in situations where the lack of a
direction in an enum direction8 variable means the tile a direction would
have been relative to.

The reason for using DIR8_ORIGIN rather than using direction8_invalid()
directly is to make it easier to separate "set to this tile" from "invalid
or buggy value" in the future. At the moment DIR8_ORIGIN is set to -1 via
direction8_invalid().

See patch #6393