Don't get the id from a potentially dead struct
Don't access a potentially dead city or unit to get its id number to check
if it is alive. There is no way to know what that memory contains if it is
dead.

Use the already existing actor_id and target_id when unit_perform_action()
checks that actor and target survived the Lua.

The checks in the action doer functions are redundant because of the checks
in unit_perform_action(). Drop them.

Reported by Marko Lindqvist <cazfi>

See bug #23956