Allow changing sciencebox in heavy scenarios
Allow changing the sciencebox setting before the game start in scenarios
that includes players, cities, etc.

The corner case where a change to the sciencebox setting on a heavy
scenario makes a player's amount of bulbs enough to get what is being
researched is handled in srv_ready().

See patch #6483