sandbox: add disband unit rules
A unit can be Disbandable, Undisbandable or ConsensusUndisbandable. The
consensus among the player's bureaucracy is that ConsensusUndisbandable
units never should be disbanded (or recycled). Under some governments it
has the power to enforce its will.

The unit types Battleship and Carrier are ConsensusUndisbandable. Both unit
types gets an Easter Egg in their ruleset comments: a reference to a news
paper article indicating that ConsensusUndisbandable isn't a completely
fictional concept.

Adding "Disband Unit" to the sandbox ruleset was accidentally forgotten
when "Disband Unit" was introduced in patch #6573.

Reported by Marko Lindqvist <cazfi>

See patch #6596