Ruleset control missing unit upkeep wipe
Give the ruleset control over if a unit is wiped when there aren't enough
food, gold or shields to pay for its upkeep.

The immunity granted by the Undisbandable unit type flag is still hard
coded. What actions a unit is forced to perform before it is wiped, if any,
is still hard coded.

See patch #6632