Fix upkeep disband side effect crash
An action auto performer rule caused by missing upkeep can kill other units
as collateral damage. (Example: "Explode Nuclear") A unit killed as
collateral damage can be in the list of units that is missing upkeep.
Unless the missing upkeep is solved the server could end up referring to
the unit that died as collateral damage.

Remove the collateral damage as it is killed.

See bug #24124