Fix unit removal collateral damage negative gold
A player's gold is supposed to be positive during the turn. The gold upkeep
payment starts with paying the unit gold upkeep. If the gold becomes
negative a list of units with gold upkeep is generated. A unit on the list
is removed ("sold") and its gold upkeep returned until the amount of gold
becomes acceptable ("balanced").

An action auto performer rule caused by missing upkeep can kill other units
as collateral damage. (Example: "Explode Nuclear") A unit killed as
collateral damage can be on the list of units that are missing upkeep.

Pay back the gold upkeep of collateral damage units the game considers to
sell to balance the player's gold. This avoids the collateral damage
causing negative gold.

Store the payed gold upkeep for a unit so the correct gold upkeep is payed
back. This avoids a reduction of the current amount of upkeep causing
negative gold.

See bug #24132