Trade route partner city size is known
The owner of a city can see the city externals (like size) of its trade
partners. The meta knowledge for the Building requirement type was aware of
this. The meta knowledge for the MinSize requirement type wasn't.

Centralize checking if city externals are visible in the new function
mke_can_see_city_externals(). Make it aware that a trade route reveals the
city externals to a trade partner even if the city tile aren't known and
seen. Use it for both the Building and the MinSize requirement types.

See bug #24178