clients: try the next disband if the first fails.
The function request_unit_disband() will pick the best disband action that
may, from the player's point of view, be enabled. A ruleset can enable a
disband action based on information the player don't have. This can make
request_unit_disband() ask the server to perform an illegal action.

Change request_unit_disband() so it waits for the result of its request and
tries the next (worse) action if the better action was illegal.

See patch #6786