Remove non do action unit_do_action use.
Asking to be reminded to ask what actions the unit can do to a tile doesn't
change the game world. Stop sending it in the unit_do_action packet.

Asking about actions is done by the client. A unit's action decision state
therefore belongs to it even if it is stored at the server side too. The
same is true for a unit's battle group number. It is also handled client
side.

Client owned unit state is stored in the server so it can be sent to
reconnecting clients, stored in save games and, in cases like action
decision state, written to by the server.

Expand the unit_battlegroup packet to also cover action decision state. It
is therefore used to set more than one kind of unit server side client
owned state. Rename it to unit_sscs_set.

See patch #6819