Make user meta server message a setting.
This gives it access to various setting features like the /show command and
the client side server setting GUI.

Freeciv-web has a GUI for setting the user meta server message. At the
moment it has to patch Freeciv to transfer the current metamessage value to
the client. The value of server settings are transfered to the client.

Don't remove the old metamessage command. Reasons:
* it can set a message with quotation marks
* consistency with the other meta server settings
* old scripts may rely on it
* server administrators are used to it

See patch #6876