Explicit access level for setting read / write.
This allows an individual setting to demand a custom access level to read
from or write to it.

Make each setting require the same connection access levels for reading from
or writing to it as before.

Thanks to Marko Lindqvist <cazfi> for reporting an access level change
client update problem in the first version of this patch.

See patch #6961