sandbox: require a Marketplace to enter.
Make the legality of the action "Enter Marketplace" depend on the presence
of a Marketplace in the target city rather than on if a trade route can be
established.

David Fernandez <bardo> and I talked about doing this for civ2civ3 (in
patch #5995). Do it in sandbox first so it can get a bit more testing.

See patch #7041