Remove custom illegal "Found City" explanation.
Add the ability to action not enabled explain that a city was too close to
the target tile. Have it detect "Found City" citymindist problems.

Detect when founding a city is illegal because it was founded on a foreign
tile when founding a city on an unclaimed tile is illegal.

This makes the action not enabled explanation system detect everything the
old city founding specific system could. The latter can therefore be
removed.

See patch #7051