rscompat: support partial req type changes.
Make it possible for rscompat_req_type_name_3_0() to handle situations where
some, but not all, instances of a requirement type has changed type. This is
useful in cases like when a base flag becomes an extra flag and if
diplomatic treaties becomes ruleset defined and moves out of DiplRel.

See patch #7056