sandbox: let the Leader scorch the earth.
Allow the Leader unit to do the action "Destroy City" to domestic cities.

Permit scorched earth by making it possible to completely destroy cites. Use
costs and limitations to balance it's power.

Only a Leader unit can destroy a city. It has to be in or adjacent to the
target city. This will often increase the risk that he is killed.

It is impossible to destroy foreign cities. A city must be therefore
conquered before it can be destroyed.

See patch #7204