Help: specify if target must be foreign/domestic.
The auto generated help text of a unit contains what actions it can do. Each
action has a bullet point. It contains the action's target type.

Some action can only be done to foreign targets. Some actions can only be
done to domestic targets. Some actions can be done to both. This information
is significant enough to add to the help text.

Move the action's target kind to a sub bullet point. This gives room to
specify when the target must be foreign or domestic.

See patch #7246