Remove action specific transporting anek.
The action not enabled explanation system can explain that an action can't
be done because the actor unit is transporting another unit. Demand that the
hard rule forbidding a unit from airlifting or paradropping while
transporting another unit is included in the respective action's action
enablers. This allows the action specific check to be dropped.

See patch #7327