More event graphics, and move some from small.spec.
The graphics used for e_* sprites that were in small.png [1, {0,1,2,5}]
have been moved to events.png. Most of the remaining icons on the second
row of small.png are currently unused and have been moved to the
right-hand side.

This events.png is targeted at the 2.6 event set; sprite choices for
events that are new since then are placeholders.

Event graphics and spec updates by GriffonSpade on the forum.

See gna patch #7439.