Gtk 3 clients: add the "Go to and..." menu.
Make it work like the Qt client's "Go to and..." menu for now. The user
selects the action before clicking on the map. No checks for action legality
are performed. This allows the player to order an action that will become
legal before the unit reaches the target tile. Example: join the city that
will be built at the target tile next turn.

Implementing a more beginner friendly alternative UI like the one suggested
in bug #16905 has been left for later.

See patch #7644