Freeciv-web: ruleset action distance compat.
An unlimited max action distance is stored as the text "unlimited". The
largest non unlimited max distance an action can have is stored as a
number. The max size of that number was MAP_DISTANCE_MAX. MAP_DISTANCE_MAX
is the largest possible distance between two tiles. But that value is,
because of different max map sizes, smaller in Freeciv-web than in regular
Freeciv. This made it possible to create a ruleset that would be rejected in
Freeciv-web because of a too larger action max distance.

Define the action distance numbers without a direct reference to
MAP_DISTANCE_MAX. Add a static assertion to make sure that MAP_DISTANCE_MAX
still can be as a regular distance. This makes Freeciv and Freeciv-web
rulesets compatible (when it comes to action distance).

See patch #7715