Lua move: city conquest only when possible.
The Lua functions unit_teleport() and unit_move() can result in the conquest
of a city on the tile the unit is moved to. Don't try to conquer unless the
target tile has a city, the actor unit can take over cities and the actor
unit's owner is at war with the target city's owner.

See patch #7754